Add debug task with args feature
