+++ extension/src/commands.ts
@@ -17,13 +17,8 @@ import { getConfigIsTasksExplorerEnabled } from './config';
 import { GradleTasksClient } from './client';
 import { logger } from './logger';
 import {
-  isJavaDebuggerExtensionActivated,
-  JAVA_LANGUAGE_EXTENSION_ID,
-  JAVA_DEBUGGER_EXTENSION_ID,
   isJavaLanguageSupportExtensionActivated,
   JAVA_CONFIGURATION_UPDATE_COMMAND,
-  getJavaDebuggerExtension,
-  getJavaLanguageSupportExtension,
 } from './compat';
 
 const localize = nls.loadMessageBundle();
@@ -66,37 +61,9 @@ function registerDebugTaskCommand(
 ): vscode.Disposable {
   return vscode.commands.registerCommand(
     'gradle.debugTask',
-    async (treeItem: GradleTaskTreeItem) => {
+    async (treeItem: GradleTaskTreeItem, args = '') => {
       if (treeItem && treeItem.task) {
-        const INSTALL_EXTENSIONS = localize(
-          'commands.requiredExtensionMissing',
-          'Install Missing Extensions'
-        );
-        if (!getJavaLanguageSupportExtension() || !getJavaDebuggerExtension()) {
-          const input = await vscode.window.showErrorMessage(
-            localize(
-              'commands.missingJavaLanguageSupportExtension',
-              'The Java Language Support & Debugger extensions are required for debugging.'
-            ),
-            INSTALL_EXTENSIONS
-          );
-          if (input === INSTALL_EXTENSIONS) {
-            await vscode.commands.executeCommand(
-              'workbench.extensions.action.showExtensionsWithIds',
-              [JAVA_LANGUAGE_EXTENSION_ID, JAVA_DEBUGGER_EXTENSION_ID]
-            );
-          }
-          return;
-        } else if (!isJavaDebuggerExtensionActivated()) {
-          vscode.window.showErrorMessage(
-            localize(
-              'commands.javaDebuggerExtensionNotActivated',
-              'The Java Debugger extension is not activated.'
-            )
-          );
-          return;
-        }
-        runTask(treeItem.task, client, true);
+        runTask(treeItem.task, client, args, true);
       }
     }
   );
@@ -123,7 +90,20 @@ function registerRunTaskWithArgsCommand(
     'gradle.runTaskWithArgs',
     (treeItem: GradleTaskTreeItem) => {
       if (treeItem && treeItem.task) {
-        runTaskWithArgs(treeItem.task, client);
+        runTaskWithArgs(treeItem.task, client, false);
+      }
+    }
+  );
+}
+
+function registerDebugTaskWithArgsCommand(
+  client: GradleTasksClient
+): vscode.Disposable {
+  return vscode.commands.registerCommand(
+    'gradle.debugTaskWithArgs',
+    (treeItem: GradleTaskTreeItem) => {
+      if (treeItem && treeItem.task) {
+        runTaskWithArgs(treeItem.task, client, true);
       }
     }
   );
@@ -330,6 +310,7 @@ export function registerCommands(
     registerDebugTaskCommand(client),
     registerRestartTaskCommand(),
     registerRunTaskWithArgsCommand(client),
+    registerDebugTaskWithArgsCommand(client),
     registerCancelTaskCommand(statusBarItem),
     registerCancelTreeItemTaskCommand(),
     registerRefreshCommand(taskProvider, treeDataProvider),
